requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
absl-py=0.10.0=py37hc8dfbb8_0
appdirs=1.4.3=py_1
attrs=20.1.0=pyh9f0ad1d_0
beautifulsoup4=4.9.1=py_1
black=19.10b0=py_4
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py37h60d8a13_1000
c-ares=1.16.1=haf1e3a3_3
ca-certificates=2020.7.22=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py37hc8dfbb8_0
cffi=1.14.3=py37hf5b7abd_0
chardet=3.0.4=py37hc8dfbb8_1007
click=7.1.2=pyh9f0ad1d_0
cryptography=3.1=py37h94e4008_0
cycler=0.10.0=py_2
flake8=3.8.3=py_1
freetype=2.10.2=ha233b18_0
google-auth=1.21.2=py_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.31.0=py37hbfa9a1b_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py37hc8dfbb8_0
intel-openmp=2019.4=233
jpeg=9b=he5867d9_2
kiwisolver=1.2.0=py37ha1cc60f_0
lcms2=2.11=h92f6f08_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libcxx=10.0.1=h5f48129_0
libffi=3.2.1=hb1e8313_1007
libgfortran=4.0.0=2
liblapack=3.8.0=14_mkl
libmklml=2019.0.5=0
libopenblas=0.3.10=openmp_h63d9170_4
libpng=1.6.37=ha441bb4_0
libprotobuf=3.13.0=hab81aa3_0
libtiff=4.1.0=hcb84e12_1
llvm-openmp=10.0.1=h28b9765_0
lz4-c=1.9.2=hb1e8313_1
markdown=3.2.2=py_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h886f89f_0
mccabe=0.6.1=py_1
mkl=2019.5=281
mkl-service=2.3.0=py37h0b31af3_0
ncurses=6.2=hb1e8313_1
ninja=1.10.1=h879752b_1
numpy=1.19.1=py37h1efc2f6_2
oauthlib=3.0.1=py_0
olefile=0.46=py_0
openssl=1.1.1g=haf1e3a3_1
pandas=1.1.2=py37hdadc0f0_0
pathspec=0.8.0=pyh9f0ad1d_0
pillow=7.2.0=py37ha54b6ba_0
pip=20.2.2=py_0
protobuf=3.13.0=py37hdadc0f0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=pyh9f0ad1d_0
pycparser=2.20=pyh9f0ad1d_2
pyflakes=2.2.0=pyh9f0ad1d_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.8=hc9dea61_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.5.1=py3.7_0
pytz=2020.1=pyh9f0ad1d_0
readline=8.0=h0678c8f_2
regex=2020.7.14=py37h60d8a13_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
setuptools=49.6.0=py37hc8dfbb8_1
six=1.15.0=pyh9f0ad1d_0
soupsieve=2.0.1=py_1
sqlite=3.32.3=h960bd1c_1
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.1=pyh9f0ad1d_0
torchvision=0.6.1=py37_cpu
tornado=6.0.4=py37h9bfed18_1
tqdm=4.49.0=pyh9f0ad1d_0
typed-ast=1.4.1=py37h0b31af3_0
urllib3=1.25.10=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py_1
xz=5.2.5=haf1e3a3_1
zipp=3.1.0=py_0
zlib=1.2.11=1007
zstd=1.4.4=h1990bb4_3
